Added FetchRequests and Intents
Added FetchRequest for SproutCareInformationMO
Added Intents for SproutCareInformationMO
Added Intents for SproutPlantMO
